Batch rename of _EVERYTHING_
