+++ README.md
@@ -1,4 +1,4 @@
-AtoMS - Simple Microservice for Data Units
+IteMS - Simple Microservice for Data Units
 ==========================================
 
 A simple microservice built on top of Spark (sparkjava.com) for holding small
@@ -14,22 +14,22 @@ Run
 ---
 To run the microservice without any persistence:
 
-$ java -jar build/libs/atoms.jar
+$ java -jar build/libs/items.jar
 
 To run the microservice backed by a MapDB map:
 
-$ java -jar build/libs/atoms.jar --mapdb --mapdb-file atoms.db --mapdb-map atoms
+$ java -jar build/libs/items.jar --mapdb --mapdb-file items.db --mapdb-map items
 
 To run the microservice backed by a MongoDB collection:
 
-$ java -jar build/libs/atoms.jar --mongodb --mongdb-host localhost \
-    --mongodb-port 27017 --mongo-db atoms --mongodb-col test-collection
+$ java -jar build/libs/items.jar --mongodb --mongdb-host localhost \
+    --mongodb-port 27017 --mongo-db items --mongodb-col test-collection
 
 Use
 ---
 To use the service, access the URL:
 
-http://localhost:9999/atoms/[name]
+http://localhost:9999/items/[name]
 
 where [name] is the data you want to access.
