--- atoms
@@ -1,45 +1,0 @@
-#!/bin/bash
-# Atoms startup script
-
-# If not set, set ATOMS_HOME to current dir
-ATOMS_HOME="${ATOMS_HOME:-$(readlink -m "$(dirname $0)")}"
-
-# Set the defaults
-CONFIG_FILE="${ATOMS_HOME}/etc/atoms.conf"
-ATOMS_JAR="${ATOMS_HOME}/build/libs/atoms.jar"
-ATOMS_ARGS=""
-
-# Read CLI params
-while [ -n "$1" ]
-do
-    case $1 in
-        --conf) 
-            CONFIG_FILE="$2"
-            shift
-            ;;
-        --jar)
-            ATOMS_JAR="$2"
-            shift
-            ;;
-        *)
-            ATOMS_ARGS="${ATOM_ARGS} $1"
-            ;;
-    esac
-    shift
-done
-
-if [ -f "${CONFIG_FILE}" ]
-then
-    echo "Loading config file ${CONFIG_FILE}"
-    . "${CONFIG_FILE}"
-else
-    echo "Unable to find config file ${CONFIG_FILE}" 1>&2
-fi
-
-if [ ! -f "${ATOMS_JAR}" ]
-then
-    echo "Unable to find atoms in ${ATOMS_JAR}" 1>&2
-    exit 1
-fi
-
-exec java ${JAVA_OPTS} -jar ${ATOMS_JAR} ${ATOMS_OPTS} ${ATOMS_ARGS}
