--- etc/atoms.conf
@@ -1,4 +1,0 @@
-#
-# Atoms config file
-
-ATOMS_OPTS="--mapdb --mapdb-file ${ATOMS_HOME}/var/atoms.db --mapdb-map atoms"
