+++ etc/items.conf
@@ -1,0 +1,4 @@
+#
+# Default items config file
+
+ITEMS_OPTS="--mapdb --mapdb-file ${ITEMS_HOME}/var/items.db --mapdb-map items"
