+++ items
@@ -1,0 +1,62 @@
+#!/bin/bash
+# Copyright 2016, 2019, Matt Dean
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+#
+# Items startup script
+#
+
+# If not set, set ITEMS_HOME to current dir
+ITEMS_HOME="${ITEMS_HOME:-$(readlink -m "$(dirname $0)")}"
+
+# Set the defaults
+CONFIG_FILE="${ITEMS_HOME}/etc/items.conf"
+ITEMS_JAR="${ITEMS_HOME}/build/libs/items.jar"
+ITEMS_ARGS=""
+
+# Read CLI params
+while [ -n "$1" ]
+do
+    case $1 in
+        --conf) 
+            CONFIG_FILE="$2"
+            shift
+            ;;
+        --jar)
+            ITEMS_JAR="$2"
+            shift
+            ;;
+        *)
+            ITEMS_ARGS="${ITEMS_ARGS} $1"
+            ;;
+    esac
+    shift
+done
+
+if [ -f "${CONFIG_FILE}" ]
+then
+    echo "Loading config file ${CONFIG_FILE}"
+    . "${CONFIG_FILE}"
+else
+    echo "Unable to find config file ${CONFIG_FILE}" 1>&2
+fi
+
+if [ ! -f "${ITEMS_JAR}" ]
+then
+    echo "Unable to find items in ${ITEMS_JAR}" 1>&2
+    exit 1
+fi
+
+exec java ${JAVA_OPTS} -jar ${ITEMS_JAR} ${ITEMS_OPTS} ${ITEMS_ARGS}
