+++ src/main/java/org/oddcyb/items/ItemsClient.java
@@ -1,7 +1,19 @@
 /*
+ * Copyright 2018, 2019, Matt Dean
  *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-package org.oddcyb.atoms;
+package org.oddcyb.items;
 
 import com.google.gson.Gson;
 import com.mashape.unirest.http.HttpResponse;
@@ -17,14 +29,14 @@ import java.util.logging.Logger;
 /**
  *
  */
-public class AtomsClient 
+public class ItemsClient 
 {
     private static final Logger LOG = 
-        Logger.getLogger(AtomsClient.class.getName());
+        Logger.getLogger(ItemsClient.class.getName());
 
     private final String url;
     
-    public AtomsClient(String url)
+    public ItemsClient(String url)
     {
         this.url = url;
     }
