+++ src/main/java/org/oddcyb/items/handlers/StoreGet.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Matt Dean
+ * Copyright 2018, 2019, Matt Dean
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.oddcyb.atoms.handlers;
+package org.oddcyb.items.handlers;
 
 import com.google.gson.Gson;
-import org.oddcyb.atoms.store.Store;
-import spark.Redirect;
+import org.oddcyb.items.store.Store;
 import spark.Request;
 import spark.Response;
 import spark.Route;
