+++ src/main/java/org/oddcyb/items/handlers/StoreReplace.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016, 2017 Matt Dean
+ * Copyright 2016, 2019, Matt Dean
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.oddcyb.atoms.handlers;
+package org.oddcyb.items.handlers;
 
 import com.google.gson.Gson;
-import org.oddcyb.atoms.store.Store;
+import org.oddcyb.items.store.Store;
 import spark.Request;
 import spark.Response;
 import spark.Route;
@@ -24,37 +24,24 @@ import spark.Route;
 /**
  *
  */
-public class StoreAdd implements Route
+public class StoreReplace implements Route
 {
-    public static final String SUCCESS_RESPONSE = 
-        "<html><body>201 Created</body></html>";
 
     private final Store store;
     private final Gson gson = new Gson();
     
-    public StoreAdd(Store store)
+    public StoreReplace(Store store)
     {
         this.store = store;
     }
-    
+
     @Override
-    public Object handle(Request req, Response resp) throws Exception
+    public Object handle(Request req, Response resp) throws Exception 
     {
         String name = req.splat()[0];
         String json = req.body();
         
-        Object exists = this.store.add(name, gson.fromJson(json, Object.class));
-        
-        if ( exists == null )
-        {
-            resp.status(201);
-            return SUCCESS_RESPONSE;
-        }
-        else
-        {
-            resp.status(409);
-            return exists;
-        }
+        return this.store.replace(name, this.gson.fromJson(json, Object.class));
     }
-
+    
 }
