+++ src/main/java/org/oddcyb/items/store/MongoDBStore.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Matt Dean
+ * Copyright 2016, 2019, Matt Dean
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.oddcyb.atoms.store;
+package org.oddcyb.items.store;
 
 import com.mongodb.MongoClient;
 import com.mongodb.client.MongoCollection;
